Add display, addRefresh and apiManipulation modules
