+++ src/index.html
@@ -23,7 +23,7 @@
             <h2>Add Your Score</h2>
             <form action="/" id="input-container">
                 <input type="text" placeholder="Your Name" id="player-input"><br>
-                <input type="text" placeholder="Your score" id="score-input" required><br>
+                <input type="number" placeholder="Your score" id="score-input" required><br>
                 <button type="submit">Submit</button>
             </form>
         </div>
